requirements.txt
torch
torchvision
torchaudio
wandb
pandas
datasets
python-dotenv
huggingface-hub
numpy
tqdm
tensorboard
protobuf